fix: rename resolution utilities
