+++ packages/test-apps/react-vite/src/resolutionUtilities/index.ts
@@ -1,2 +1,2 @@
-export * from './basic';
+export * from './typeGuards';
 export * from './resolution';
