+++ packages/test-apps/react-vite/src/resolutionUtilities/typeGuards.ts
